develop of new commands
